Reimplement source tree scanning with File::Find. This both simplifies the code and makes it easier to extend Nancy to handle multiple trees.
Add a FIXME about merging the source file map and leaf check.

Improve formatting of debug output.

Fix formatting of help message, and reorder options in Getopt call to
be the same as in the help message.

Return the empty string if we can't find an $include file. The
existing code had the same effect, but only because undef was
coerced by s/// to the empty string.