Various fixes to the new features of Nancy:
Add $find command (currently undocumented, only used in one place). It
works like $include but returns the name of the file rather than its
contents. It is used to find the XML file when converting XML to HTML
in multi-tree builds.

Rename $object to $obj for consistency with makesite.pl.

It's tar that has an --exclude-vcs option, not find.

Warn if we can't write a file, don't die, so the rest of the site can
be built.

Fix ordering of file finding: we should try each tree at a given
level, not all the levels within each tree, so that one tree can
override a file at a given level in another.

Fix emptyDir to take account of .svn directories.