Calls to internal methods do not need WWW::Nancy:: prefixes.
Add tree_dump to aid debugging.

Make findFragment work with fragment names that are paths, and in
particular, that include . and .. components.

Fix a bug when warnings not turned on would make wrong fragment be
fetched by findFragment.

Speed up Nancy by storing file names, not contents, in the tree, and
only reading contents when actually needed.