Rejig how files are added for output by expand_tree: add a method add_output to add a file, and only write files out that are explicitly added.
Fix a bug in tree_set and tree_delete, which both modified their $path
argument.

As a result of the rejig, change the code that reads the sources in
makesite.pl to be the same as in webnancy_adsensus.pl, as it can read
the site framework before doing the merge.

Unfortunately, files that are only linked to are still not added to
the final site, so some links will be broken.