Implement lazy tree merging in order to simplify link processing.
As a result, it was possible to stop storing input and output files in
intermediate data structures. This makes Nancy rather slower, but not
disastrously slow, given that we are moving to dynamic page generation
as the default. Because makesite.pl now calls nancy.pl, move the
whole-site routines back into naancy.pl. Lots of general code, naming
and comment clean-up.

Update manual to match reality.