nancy: various minor improvements.
Remove normalize_path; it’s not a big security help. Better use a
chroot.

Rename $list_files to $ListFiles for consistency with other
configuration variables, and remove note to implement it with an
environment variable, as we already have.

Accept a file even if it’s empty: we no longer use empty files for
masking.