Make Nancy chomp its $included inputs
This almost always improves the output; in particular, it makes it
easier to use the results of includes as input to other includes, and
avoids problems when an $include is in the middle of a line.

Note this behaviour in the manual; also update the documentation of Perl
dependencies and stop using my custom copyrght [sic!] package.

Also simplify the way Nancy deals with the special filename `-': in
principle allow it anywhere, though in practice it can only be used
once. Not currently documented.