Replace linkfs with hand-written merge code
This is necessary so that we can eventually provide the filesystem path of
the file currently being expanded (next commit).

Also improve some tests so that they clean up after themselves better in
case of failure and/or use temporary directories instead of the development
tree.

Improve some wording in README.md slightly.