Fix findObject’s not to return objects that should be unseen
findObject’s documentation was incorrect: for a directory, it correctly
returned its contents as an array, not a map.

However, the array contained objects which had the same name relative to the
root for example, if nancy was run with roots ‘a’ and ‘b’, and each
directory contained a file ‘foo’, findObject would return both of them. It
should only return the one under the left-most root!

There was a small clue in the code: the roots were scanned right-to-left.
This is a correct way to do it: we scan the roots right-to-left and store
the results in a map, so that results under roots to the left of the list
override results under roots to the right.