Add $outputpath, add .in.nancy support, and stricter order of processing
Sort filenames lexically, and deal with them in three groups. This is
motivated by wanting `.in.nancy` and $outputpath, so we can have a templated
file that is expanded purely for its side effects, which can inlude writing
files to the output.

Document, and add a test and cookbook example.