Rework the way external commands are called
Rather than simply run, rather than read, a file found to be executable, add
new “external arguments”. Both `$paste` and `$include` can now have regular
and/or external arguments. When external arguments are provided, they
specify a program and command-line arguments to run. The program is looked
up as before, so it can be found via the include mechanism, or on PATH.

When an input is provided, feed it to the program on standard input.

As before, the command expands to the program’s standard output.