Separate `scandir()` from `find_object()`.
This is a change of semantics in a rare case: specifically, if the same
object appears in multiple input trees and is a directory in the
leftmost input tree but not in the others. The old behaviour was to use
the leftmost non-directory. The new behaviour is to use the directory.